requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
streamlit==1.28.1
python-docx==0.8.11
PyPDF2==3.0.1
tiktoken==0.5.1
openai==1.3.5